A......... [ZBXNEXT-1575] removed host prototype support for discovered hosts ..F....... [ZBXNEXT-1575] hostprototype.get: will no longer return prototypes from discovered hosts ..F....... [ZBXNEXT-1575] hostprototype.create: it is no longer possible to create host prototypes on discovered hosts
git-svn-id: svn://svn.zabbix.com/branches/dev@34651 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBXNEXT-1633/frontends/php/api/classes/CHostPrototype.php
@@ -650,7 +650,9 @@ class CHostPrototype extends CHostBase {
 		$discoveredHosts = DBfetchArray(DBselect(
 			'SELECT hostid FROM host_discovery WHERE '.dbConditionInt('parent_hostid', $hostPrototypeIds)
 		));
-		API::Host()->delete(zbx_objectValues($discoveredHosts, 'hostid'));
+		if ($discoveredHosts) {
+			API::Host()->delete(zbx_objectValues($discoveredHosts, 'hostid'));
+		}
 
 		// delete host prototypes
 		DB::delete($this->tableName(), array('hostid' => $hostPrototypeIds));
